* develop:
  apply style to recommendation card in timeline
  apply styel to app update card
  add store name and avatar to app update in timeline
  remove timeline fake update
  migrate download logic to install manager
  add view attached and detached life cycle events to widget
  remove unecessary code for root install
  install manager attempts to install using root
  add support for unbind method in active and completed download widgets
  fix default install observable making sure it only emits an item when installation is finished
  use unbind method in update widget to unsubscribe from observable
  make as few changes as possible in app install widget to avoid conflicts
  add unbind methods to widget class and call it from base adapter
  implement silent install with fallback to default install
  replace static install and uninstall methods for install manager class
  corrected the bug that was thrown when resuming a download
  always emit all download states

# Conflicts:
#	v8engine/src/main/res/values/colors.xml

+++ downloadmanager/src/main/java/cm/aptoide/pt/downloadmanager/AptoideDownloadManager.java
@@ -14,7 +14,7 @@ import com.liulishuo.filedownloader.FileDownloader;
 import java.util.ArrayList;
 import java.util.List;
 
-import cm.aptoide.pt.actions.PermissionRequest;
+import cm.aptoide.pt.actions.PermissionManager;
 import cm.aptoide.pt.database.Database;
 import cm.aptoide.pt.database.realm.Download;
 import cm.aptoide.pt.database.realm.FileToDownload;
@@ -28,7 +28,6 @@ import io.realm.Sort;
 import lombok.AccessLevel;
 import lombok.Cleanup;
 import lombok.Getter;
-import lombok.Setter;
 import rx.Observable;
 
 /**
@@ -55,8 +54,6 @@ public class AptoideDownloadManager {
 	private boolean isDownloading = false;
 	@Getter(AccessLevel.MODULE) private DownloadNotificationActionsInterface downloadNotificationActionsInterface;
 	@Getter(AccessLevel.MODULE) private DownloadSettingsInterface settingsInterface;
-	@Setter
-	private PermissionRequest permissionRequest;
 
 	public static Context getContext() {
 		return context;
@@ -67,16 +64,6 @@ public class AptoideDownloadManager {
 			instance = new AptoideDownloadManager();
 		}
 		return instance;
-	}
-
-	private static Observable<Download> getDownload(Realm realm, long appId) {
-		return Observable.fromCallable(() -> realm.where(Download.class).equalTo("appId", appId).findFirst()).flatMap(download -> {
-			if (download == null || (download.getOverallDownloadStatus() == Download.COMPLETED && getInstance().getStateIfFileExists(download) == Download
-					.FILE_MISSING)) {
-				return Observable.error(new DownloadNotFoundException());
-			}
-			return download.asObservable();
-		});
 	}
 
 	void initDownloadService(Context context) {
@@ -103,15 +90,7 @@ public class AptoideDownloadManager {
 		if (getDownloadStatus(download.getAppId()) == Download.COMPLETED) {
 			return Observable.fromCallable(() -> download);
 		}
-		if (permissionRequest != null) {
-			permissionRequest.requestAccessToExternalFileSystem(() -> {
-				startNewDownload(download);
-			}, ()-> {
-				Logger.e(TAG, "unable to start new download");
-			});
-		} else {
-			startNewDownload(download);
-		}
+		startNewDownload(download);
 		return getDownload(download.getAppId());
 	}
 
@@ -127,7 +106,7 @@ public class AptoideDownloadManager {
 
 	public void pauseDownload(long appId) {
 		@Cleanup Realm realm = Database.get();
-		Download download = getDownloadObject(appId);
+		Download download = getStoredDownload(appId, realm);
 		if (download != null) {
 			if (download.getOverallDownloadStatus() != Download.PROGRESS) {
 				download.setOverallDownloadStatus(Download.PAUSED);
@@ -147,18 +126,22 @@ public class AptoideDownloadManager {
 	 * @return observable for download state changes.
 	 */
 	public Observable<Download> getDownload(long appId) {
-		@Cleanup
 		Realm realm = Database.get();
-		return getDownload(realm, appId).takeUntil(download -> download.getOverallDownloadStatus() != Download.PROGRESS && download.getOverallDownloadStatus()
-				!= Download.IN_QUEUE && download
-				.getOverallDownloadStatus() != Download.PENDING && download.getOverallDownloadStatus() != Download.INVALID_STATUS).map(Download::clone);
+		Download download = realm.where(Download.class).equalTo("appId", appId).findFirst();
+		if (download == null ||
+				(download.getOverallDownloadStatus() == Download.COMPLETED && getInstance().getStateIfFileExists(download) == Download.FILE_MISSING)) {
+			realm.close();
+			return Observable.error(new DownloadNotFoundException());
+		} else {
+			return download.<Download>asObservable().map(Download::clone).takeUntil(storedDownload -> storedDownload.getOverallDownloadStatus() == Download.COMPLETED);
+		}
 	}
 
-	Observable<Download> getCurrentDownload() {
+	public Observable<Download> getCurrentDownload() {
 		return getDownloads().flatMapIterable(downloads -> downloads).filter(downloads -> downloads.getOverallDownloadStatus() == Download.PROGRESS);
 	}
 
-	Observable<List<Download>> getCurrentDownloads() {
+	public Observable<List<Download>> getCurrentDownloads() {
 		@Cleanup
 		Realm realm = Database.get();
 		RealmResults<Download> download = realm.where(Download.class).findAll();
@@ -178,21 +161,16 @@ public class AptoideDownloadManager {
 		}
 	}
 
-	Observable<List<Download>> getDownloads() {
+	public Observable<List<Download>> getDownloads() {
 		@Cleanup
 		Realm realm = Database.get();
-		RealmResults<Download> download = realm.where(Download.class).findAll();
-		if (download.size() >= 0) {
-			return download.asObservable().map(results -> {
-				List<Download> list = new ArrayList<>();
-				for (final Download result : results) {
-					list.add(result.clone());
-				}
-				return list;
-			});
-		} else {
-			return Observable.error(new DownloadNotFoundException());
-		}
+		return realm.where(Download.class).findAll().asObservable().map(results -> {
+			List<Download> list = new ArrayList<>();
+			for (final Download result : results) {
+				list.add(result.clone());
+			}
+			return list;
+		});
 	}
 
 	/**
@@ -200,6 +178,7 @@ public class AptoideDownloadManager {
 	 */
 	public void pauseAllDownloads() {
 		FileDownloader.getImpl().pauseAll();
+
 		getCurrentDownloads().first().subscribe(downloads -> {
 			@Cleanup
 			Realm realm = Database.get();
@@ -214,7 +193,8 @@ public class AptoideDownloadManager {
 	}
 
 	private int getDownloadStatus(long appId) {
-		Download download = getDownloadObject(appId);
+		@Cleanup Realm realm = Database.get();
+		Download download = getStoredDownload(appId, realm);
 		if (download != null) {
 			if (download.getOverallDownloadStatus() == Download.COMPLETED) {
 				return getStateIfFileExists(download);
@@ -237,13 +217,13 @@ public class AptoideDownloadManager {
 		if (TextUtils.isEmpty(OBB_PATH)) {
 			OBB_PATH = GENERIC_PATH;
 		}
-
 	}
 
 	@NonNull
 	@Download.DownloadState
-	int getStateIfFileExists(Download downloadToCheck) {
-		@Download.DownloadState int downloadStatus = Download.COMPLETED;
+	private int getStateIfFileExists(Download downloadToCheck) {
+		@Download.DownloadState
+		int downloadStatus = Download.COMPLETED;
 		for (final FileToDownload fileToDownload : downloadToCheck.getFilesToDownload()) {
 			if (!FileUtils.fileExists(fileToDownload.getFilePath())) {
 				downloadStatus = Download.FILE_MISSING;
@@ -297,21 +277,21 @@ public class AptoideDownloadManager {
 
 	public void removeDownload(long appId) {
 		@Cleanup
-		Realm realm1 = Database.get();
-		getDownload(realm1, appId).subscribe(download -> {
+		Realm realm = Database.get();
+		Download download = realm.where(Download.class).equalTo("appId", appId).findFirst();
+		if (download != null) {
 			for (final FileToDownload fileToDownload : download.getFilesToDownload()) {
 				FileUtils.removeFile(fileToDownload.getFilePath());
 			}
-			@Cleanup
-			Realm realm = Database.get();
 			Database.delete(download, realm);
-		}, Throwable::printStackTrace);
+		}
 	}
 
-	public Download getDownloadObject(long appId) {
-		@Cleanup
-		Realm realm = Database.get();
+	public Download getStoredDownload(long appId, Realm realm) {
 		Download download = realm.where(Download.class).equalTo("appId", appId).findFirst();
-		return download != null ? download.clone() : null;
+		if (download != null) {
+			return download.clone();
+		}
+		return null;
 	}
 }
